Explicitly open and close message context
