+++ src/transport/CMakeLists.txt
@@ -4,17 +4,17 @@
 
 set(HEADERS
     "${FAABRIC_INCLUDE_DIR}/faabric/transport/common.h"
+    "${FAABRIC_INCLUDE_DIR}/faabric/transport/context.h"
     "${FAABRIC_INCLUDE_DIR}/faabric/transport/macros.h"
     "${FAABRIC_INCLUDE_DIR}/faabric/transport/Message.h"
-    "${FAABRIC_INCLUDE_DIR}/faabric/transport/MessageContext.h"
     "${FAABRIC_INCLUDE_DIR}/faabric/transport/MessageEndpoint.h"
     "${FAABRIC_INCLUDE_DIR}/faabric/transport/MessageEndpointServer.h"
     "${FAABRIC_INCLUDE_DIR}/faabric/transport/MpiMessageEndpoint.h"
 )
 
 set(LIB_FILES
+    context.cpp
     Message.cpp
-    MessageContext.cpp
     MessageEndpoint.cpp
     MessageEndpointServer.cpp
     MpiMessageEndpoint.cpp
